+++ front/app/hooks/useProject.ts
@@ -30,8 +30,12 @@ export default function useProject({ projectId, projectSlug }: Props) {
     }
 
     const subscription = observable.subscribe((response) => {
-      const project = !isNilOrError(response) ? response.data : null;
-      setProject(project);
+      if (isNilOrError(response)) {
+        setProject(response);
+        return;
+      }
+
+      setProject(response.data);
     });
 
     return () => subscription.unsubscribe();