+++ front/app/modules/commercial/insights/hooks/useInsightsInputs.test.ts
@@ -214,8 +214,9 @@ describe('useInsightsInputs', () => {
   it('should return error when error', () => {
     const error = new Error();
     mockObservable = new Observable((subscriber) => {
-      subscriber.next({ data: new Error() });
+      subscriber.next(error);
     });
+
     const { result } = renderHook(() => useInsightsInputs(viewId));
     expect(result.current.list).toStrictEqual(error);
     expect(result.current.loading).toStrictEqual(false);
@@ -223,8 +224,9 @@ describe('useInsightsInputs', () => {
   });
   it('should return null when data is null', () => {
     mockObservable = new Observable((subscriber) => {
-      subscriber.next({ data: null });
+      subscriber.next(null);
     });
+
     const { result } = renderHook(() => useInsightsInputs(viewId));
     expect(result.current.list).toBe(null);
     expect(result.current.loading).toStrictEqual(false);