+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## Next
+
+## 2021-12-16
+
+### Added
+
+- Ability to scan all post, recently added posts and not tagged posts in Insights
+
+## 2021-12-15
+
+### Fixed
+
+- Severe code-injection vulnerability
+- More small copy changes for customizable navbar, made styling Navigation tab consistent with other tabs, re-enabled slug editing on secret pages-page.
+
 ## 2021-12-10
 
 - Copy for customizable navbar